Create package middleware
