+++ cmd/root.go
@@ -2,17 +2,14 @@ package cmd
 
 import (
 	"fmt"
-	"net/http"
 	"os"
 
-	"github.com/auth0/go-jwt-middleware"
-	"github.com/dgrijalva/jwt-go"
 	"github.com/gorilla/mux"
 	// postgres driver for migrate
 	_ "github.com/mattes/migrate/driver/postgres"
-	"github.com/nuveo/prest/adapters/postgres"
 	"github.com/nuveo/prest/config"
 	"github.com/nuveo/prest/controllers"
+	"github.com/nuveo/prest/middlewares"
 	"github.com/spf13/cobra"
 	"github.com/urfave/negroni"
 )
@@ -43,53 +40,15 @@ func init() {
 	config.InitConf()
 }
 
-func handlerSet(w http.ResponseWriter, r *http.Request, next http.HandlerFunc) {
-	w.Header().Set("Content-Type", "application/json")
-	next(w, r)
-}
-
-// AccessControl is a middleware to handle permissions on tables in pREST
-func AccessControl(rw http.ResponseWriter, rq *http.Request, next http.HandlerFunc) {
-	mapPath := getVars(rq.URL.Path)
-	if mapPath == nil {
-		next(rw, rq)
-		return
-	}
-
-	permission := permissionByMethod(rq.Method)
-	if permission == "" {
-		next(rw, rq)
-		return
-	}
-
-	if postgres.TablePermissions(mapPath["table"], permission) {
-		next(rw, rq)
-		return
-	}
-
-	err := fmt.Errorf("required authorization to table %s", mapPath["table"])
-	http.Error(rw, err.Error(), http.StatusUnauthorized)
-}
-
-func jwtMiddleware(key string) negroni.Handler {
-	jwtMiddleware := jwtmiddleware.New(jwtmiddleware.Options{
-		ValidationKeyGetter: func(token *jwt.Token) (interface{}, error) {
-			return []byte(key), nil
-		},
-		SigningMethod: jwt.SigningMethodHS256,
-	})
-	return negroni.HandlerFunc(jwtMiddleware.HandlerWithNext)
-}
-
 func app() {
 	cfg := config.Prest{}
 	config.Parse(&cfg)
 
 	n := negroni.Classic()
-	n.Use(negroni.HandlerFunc(handlerSet))
-	n.Use(negroni.HandlerFunc(AccessControl))
+	n.Use(negroni.HandlerFunc(middlewares.HandlerSet))
+	n.Use(negroni.HandlerFunc(middlewares.AccessControl))
 	if cfg.JWTKey != "" {
-		n.Use(jwtMiddleware(cfg.JWTKey))
+		n.Use(middlewares.JwtMiddleware(cfg.JWTKey))
 	}
 	r := mux.NewRouter()
 	r.HandleFunc("/databases", controllers.GetDatabases).Methods("GET")
